fix: spid_cie_oidc.provider.views divided in different files
