+++ spid_cie_oidc/onboarding/views.py
@@ -111,6 +111,7 @@ def onboarding_convert_jwk(request):
             return render(request, 'onboarding_convert_jwk.html', context)
     return render(request, 'onboarding_convert_jwk.html', context)
 
+
 def onboarding_convert_pem(request):
     pem_type = request.GET.get('type')
     context = {
