+++ spid_cie_oidc/provider/views/authz_request_view.py
@@ -13,19 +13,14 @@ from django.http import (
 )
 from django.shortcuts import render
 from django.urls import reverse
-from django.utils import timezone
 from django.utils.translation import gettext as _
 from django.views import View
 from pydantic import ValidationError as pydantic_ValidationError
 from spid_cie_oidc.entity.exceptions import InvalidEntityConfiguration
-from spid_cie_oidc.entity.models import (
-    TrustChain
-)
-from spid_cie_oidc.entity.tests.settings import *
-from spid_cie_oidc.provider.models import IssuedToken, OidcSession
+from spid_cie_oidc.provider.forms import AuthLoginForm, AuthzHiddenForm
+from spid_cie_oidc.provider.models import OidcSession
 
 from spid_cie_oidc.provider.exceptions import AuthzRequestReplay
-from spid_cie_oidc.provider.forms import *
 
 from . import OpBase
 logger = logging.getLogger(__name__)
@@ -224,71 +219,3 @@ class AuthzRequestView(OpBase, View):
                 logger.error(f"testigng page url reverse failed: {e}")
 
         return HttpResponseRedirect(url)
-
-class ConsentPageView(OpBase, View):
-
-    template = "op_user_consent.html"
-
-    def get_consent_form(self):
-        return ConsentPageForm
-
-    def get(self, request, *args, **kwargs):
-        try:
-            session = self.check_session(request)
-        except Exception:
-            logger.warning("Invalid session on Consent page")
-            return HttpResponseForbidden()
-
-        tc = TrustChain.objects.filter(
-            sub=session.client_id,
-            type="openid_relying_party",
-            is_active = True
-        ).first()
-
-        # if this auth code was already been used ... forbidden
-        if IssuedToken.objects.filter(session=session):
-            logger.warning(f"Auth code Replay {session}")
-            return HttpResponseForbidden()
-
-        i18n_user_claims = self.attributes_names_to_release(
-            request, session
-        )['i18n_user_claims']
-
-        context = {
-            "form": self.get_consent_form()(),
-            "session": session,
-            "client_organization_name": tc.metadata.get(
-                "client_name", session.client_id
-            ),
-            "user_claims": sorted(set(i18n_user_claims),)
-        }
-        return render(request, self.template, context)
-
-    def post(self, request, *args, **kwargs):
-        try:
-            session = self.check_session(request)
-        except Exception:
-            logger.warning("Invalid session")
-            return HttpResponseForbidden()
-
-        self.payload = session.authz_request
-        form = self.get_consent_form()(request.POST)
-        if not form.is_valid():
-            return self.redirect_response_data(
-                self.payload["redirect_uri"],
-                # TODO: this is not normative -> check AgID/IPZS
-                error="rejected_by_user",
-                error_description=_("User rejected the release of attributes"),
-                state = self.payload.get("state", "")
-            )
-        issuer = self.get_issuer()
-
-        iss_token_data = self.get_iss_token_data(session, issuer)
-        IssuedToken.objects.create(**iss_token_data)
-
-        return self.redirect_response_data(
-            self.payload["redirect_uri"],
-            code=session.auth_code,
-            state=session.authz_request["state"],
-            iss=issuer.sub if issuer else "",
-        )
